+++ binobj/fields/containers.py
@@ -33,7 +33,7 @@ class Array(Field):
         to avoid having to pass in a custom function every time.
 
     .. versionchanged:: 0.6.1
-        :meth:`.dump` and :meth:`.dumps` throw an :class:`~.errors.ArraySizeError`
+        :meth:`.to_stream` and :meth:`.to_bytes` throw an :class:`~.errors.ArraySizeError`
         if ``count`` is set and the iterable passed in is too long. It used to
         be ignored when dumping.
     """
@@ -113,14 +113,15 @@ class Array(Field):
             The sequence being checked.
         :param io.BufferedIOBase stream:
             The data stream to read from. Except in rare circumstances, this is
-            the same stream that was passed to :meth:`.load`. The stream pointer
-            should be returned to its original position when the function exits.
+            the same stream that was passed to :meth:`.from_stream`. The stream
+            pointer should be returned to its original position when the function
+            exits.
         :param list values:
             A list of the objects that have been deserialized so far. In general
             this function *should not* modify the list. A possible exception to
             this rule is to remove a sentinel value from the end of the list.
         :param context:
-            The ``context`` object passed to :meth:`.load`.
+            The ``context`` object passed to :meth:`.from_stream`.
         :param dict loaded_fields:
             The fields in the struct that have been loaded so far.
 
@@ -163,8 +164,7 @@ class Array(Field):
                 field=self, n_expected=n_elems, n_given=len(data))
 
         for value in iter(data):
-            self.component.dump(stream, value, context=context,
-                                all_fields=all_fields)
+            self.component.to_stream(stream, value, context, all_fields)
 
     def _dump_unsized(self, stream, data, n_elems, context, all_fields):
         """Dump an unsized iterable into the stream."""
@@ -177,7 +177,7 @@ class Array(Field):
                 raise errors.ArraySizeError(
                     field=self, n_expected=n_elems, n_given=n_written + 1)
 
-            self.component.dump(
+            self.component.to_stream(
                 stream, value, context=context, all_fields=all_fields)
             n_written += 1
 
@@ -203,8 +203,7 @@ class Array(Field):
         result = []
         while not self.halt_check(self, stream, result, context=context,
                                   loaded_fields=loaded_fields):
-            component = self.component.load(stream, context=context,
-                                            loaded_fields=loaded_fields)
+            component = self.component.from_stream(stream, context, loaded_fields)
             result.append(component)
 
         return result
@@ -308,7 +307,7 @@ class Union(Field):
     def _do_dump(self, stream, data, context, all_fields):
         dumper = self.dump_decider(data, self.choices, context, all_fields)
         if isinstance(dumper, Field):
-            return dumper.dump(stream, data, context, all_fields)
+            return dumper.to_stream(stream, data, context, all_fields)
 
         # Else: Dumper is not a Field instance, assume this is a Struct.
         return dumper(**data).to_stream(stream, context)
@@ -316,7 +315,7 @@ class Union(Field):
     def _do_load(self, stream, context, loaded_fields):
         loader = self.load_decider(stream, self.choices, context, loaded_fields)
         if isinstance(loader, Field):
-            return loader.load(stream, context, loaded_fields)
+            return loader.from_stream(stream, context, loaded_fields)
 
         # Else: loader is not a Field instance, assume this is a Struct.
         return loader.from_stream(stream, context)