+++ tests/fields/numeric_test.py
@@ -18,7 +18,7 @@ def test_integer_overflow():
     """An overflow error should be rethrown as a serialization error."""
     field = numeric.UInt16()
     with pytest.raises(errors.UnserializableValueError):
-        field.dumps(65536)
+        field.to_bytes(65536)
 
 
 def test_const_set_size__varint():
@@ -40,7 +40,7 @@ def test_varint__underflow():
     """Crash if VLQ gets a negative number."""
     field = numeric.VariableLengthInteger(vli_format=varints.VarIntEncoding.VLQ)
     with pytest.raises(errors.UnserializableValueError):
-        field.dumps(-1)
+        field.to_bytes(-1)
 
 
 @pytest.mark.parametrize('value,expected', (
@@ -54,7 +54,7 @@ def test_varint__basic_dump(value, expected):
     perfunctory test to make sure dumping works as expected.
     """
     field = numeric.VariableLengthInteger(vli_format=varints.VarIntEncoding.VLQ)
-    assert field.dumps(value) == expected
+    assert field.to_bytes(value) == expected
 
 
 @pytest.mark.parametrize('data, expected', (
@@ -63,7 +63,7 @@ def test_varint__basic_dump(value, expected):
 def test_varint__basic_load(data, expected):
     """Test VLQ load."""
     field = numeric.VariableLengthInteger(vli_format=varints.VarIntEncoding.VLQ)
-    assert field.loads(data) == expected
+    assert field.from_bytes(data) == expected
 
 
 def test_varint__max_bytes():
@@ -72,7 +72,7 @@ def test_varint__max_bytes():
                                           max_bytes=2)
 
     with pytest.raises(errors.ValueSizeError):
-        field.dumps(100000)
+        field.to_bytes(100000)
 
 
 def test_float_bad_endian_crashes():
@@ -94,7 +94,7 @@ def test_float_bad_endian_crashes():
     math.nan,
 ))
 def test_float__dumps(value, field_object, fmt_string):
-    assert field_object.dumps(value) == struct.pack(fmt_string, value)
+    assert field_object.to_bytes(value) == struct.pack(fmt_string, value)
 
 
 @pytest.mark.parametrize('field_object, fmt_string', (
@@ -109,20 +109,20 @@ def test_float__dumps(value, field_object, fmt_string):
     -math.inf,
 ))
 def test_float__loads(value, field_object, fmt_string):
-    assert field_object.loads(struct.pack(fmt_string, value)) == pytest.approx(value)
+    assert field_object.from_bytes(struct.pack(fmt_string, value)) == pytest.approx(value)
 
 
 @pytest.mark.skipif(_PY_VER < (3, 6), reason='binary16 only supported on 3.6+')
 def test_float16__loads():
     field = numeric.Float16(endian='little')
-    result = field.loads(struct.pack('<e', math.e))
+    result = field.from_bytes(struct.pack('<e', math.e))
     assert math.isclose(result, math.e, rel_tol=0.001)
 
 
 @pytest.mark.skipif(_PY_VER < (3, 6), reason='binary16 only supported on 3.6+')
 def test_float16__dumps():
     field = numeric.Float16(endian='big')
-    assert field.dumps(65504) == struct.pack('>e', 65504)
+    assert field.to_bytes(65504) == struct.pack('>e', 65504)
 
 
 @pytest.mark.skipif(_PY_VER >= (3, 6), reason='binary16 supported on 3.6+')
@@ -137,7 +137,7 @@ def test_float__loads__exception_translation(pack_mock):
     pack_mock.side_effect = struct.error('Some error happened')
 
     with pytest.raises(errors.DeserializationError):
-        numeric.Float32().loads(b'1234')
+        numeric.Float32().from_bytes(b'1234')
 
 
 @mock.patch('binobj.fields.numeric.struct.pack')
@@ -146,7 +146,7 @@ def test_float__dumps__exception_translation(pack_mock):
     pack_mock.side_effect = struct.error('Some error happened')
 
     with pytest.raises(errors.SerializationError):
-        numeric.Float32().dumps(1234)
+        numeric.Float32().to_bytes(1234)
 
 
 def test_timestamp__invalid_resolution():
@@ -162,19 +162,19 @@ def test_timestamp__invalid_resolution():
 ))
 def test_timestamp__loads__naive(data, expected):
     field = numeric.Timestamp32(endian='little', tz_aware=True)
-    assert field.loads(data) == expected
+    assert field.from_bytes(data) == expected
 
 
 def test_timestamp__loads__aware():
     field = numeric.Timestamp32(endian='little', tz_aware=True)
-    loaded = field.loads(b'\x01\x23\x45\x67')
+    loaded = field.from_bytes(b'\x01\x23\x45\x67')
     assert loaded == datetime.datetime(
         2024, 11, 26, 1, 23, 13, tzinfo=datetime.timezone.utc)
 
 
 def test_timestamp__loads__microseconds():
     field = numeric.Timestamp64(endian='big', resolution='us', tz_aware=True)
-    loaded = field.loads(b'\x00\x05\x81\x85\x84\x32\xc1\xad')
+    loaded = field.from_bytes(b'\x00\x05\x81\x85\x84\x32\xc1\xad')
     assert loaded == datetime.datetime(
         2019, 2, 10, 7, 55, 32, 105645, datetime.timezone.utc)
 
@@ -182,7 +182,7 @@ def test_timestamp__loads__microseconds():
 def test_timestamp__roundtrip():
     field = numeric.Timestamp(size=12, resolution='us', tz_aware=True)
     now = datetime.datetime.now(datetime.timezone.utc)
-    assert field.loads(field.dumps(now)) == now
+    assert field.from_bytes(field.to_bytes(now)) == now
 
 
 def test_timestamp__aware_cast_to_utc():
@@ -192,8 +192,8 @@ def test_timestamp__aware_cast_to_utc():
     now_est = now_utc.astimezone(est_tz)
 
     field = numeric.Timestamp64(endian='little', resolution='us', tz_aware=True)
-    assert field.dumps(now_est) == field.dumps(now_utc)
-    assert field.loads(field.dumps(now_est)) == now_utc
+    assert field.to_bytes(now_est) == field.to_bytes(now_utc)
+    assert field.from_bytes(field.to_bytes(now_est)) == now_utc
 
 
 def test_timestamp__naive_assumes_local():
@@ -203,5 +203,5 @@ def test_timestamp__naive_assumes_local():
     local_naive = local_aware.replace(tzinfo=None)
 
     field = numeric.Timestamp64(endian='little', resolution='us', tz_aware=True)
-    assert field.dumps(utc) == field.dumps(local_naive)
-    assert field.loads(field.dumps(local_naive)) == local_aware
+    assert field.to_bytes(utc) == field.to_bytes(local_naive)
+    assert field.from_bytes(field.to_bytes(local_naive)) == local_aware