+++ tests/structures_test.py
@@ -88,7 +88,7 @@ def test_load__short_read():
     """Crash if we don't have enough data to read all the fields."""
     stream = io.BytesIO(b'abcdefg\0\xba\xdc\x0f\xfe\xe1\x5b\xad\x01')
 
-    with pytest.raises(binobj.errors.UnexpectedEOFError) as errinfo:
+    with pytest.raises(errors.UnexpectedEOFError) as errinfo:
         BasicStruct.from_stream(stream)
 
     exc = errinfo.value
@@ -99,7 +99,7 @@ def test_load__short_read():
 
 def test_loads__extra_bytes():
     """Crash if we have too much data."""
-    with pytest.raises(binobj.errors.ExtraneousDataError) as errinfo:
+    with pytest.raises(errors.ExtraneousDataError) as errinfo:
         BasicStruct.from_bytes(b'abcdefghijklmnopqrstuwxyz')
 
     exc = errinfo.value
@@ -149,7 +149,7 @@ def test_partial_load__short_read_of_required():
     stream = io.BytesIO(b'zyxwvut\x0b\xad')
 
     # int64 should be the last field included in the output.
-    with pytest.raises(binobj.errors.UnexpectedEOFError):
+    with pytest.raises(errors.UnexpectedEOFError):
         BasicStruct.partial_load(stream, 'int64')
 
 
@@ -184,7 +184,7 @@ def test_dump__basic():
 ))
 def test_dump__extra_fields(extra_fields):
     """Giving unrecognized fields will crash by default."""
-    with pytest.raises(binobj.errors.UnexpectedValueError) as errinfo:
+    with pytest.raises(errors.UnexpectedValueError) as errinfo:
         BasicStruct(string='AbCdEfG', int64=-100, uint24=65535,
                     **{k: 0 for k in extra_fields})
 
@@ -195,7 +195,7 @@ def test_dump__missing_fields():
     """Crash if we're missing field a field without default values."""
     struct = BasicStruct(string='AbCdEfG', uint24=65535)
 
-    with pytest.raises(binobj.errors.MissingRequiredValueError) as errinfo:
+    with pytest.raises(errors.MissingRequiredValueError) as errinfo:
         struct.to_bytes()
 
     assert errinfo.value.field is BasicStruct.int64
@@ -209,7 +209,7 @@ class BasicStructWithDefaults(binobj.Struct):
 
 class VarlenStruct(binobj.Struct):
     n_items = fields.Int32()
-    items = fields.Array(binobj.Int32(), count=n_items)
+    items = fields.Array(fields.Int32(), count=n_items)
 
 
 def test_struct_size__basic():